pquenas coisas q faltavam. comecei a comentar o script

+++ screens/admin/script.js
@@ -1,3 +1,5 @@
+// função que lista as contas cadastradas no sistema. pega os dados do localStorage e mostra em forma de tabela
+
 function listAccounts() {
   let usuariosCadastrados = JSON.parse(localStorage.getItem("dadosCadastros"))
   let htmlListString = `<table>
@@ -43,18 +45,39 @@ function listAccounts() {
   document.getElementById('list').innerHTML = htmlListString
 }
 
+// função que exclui a conta pelo ID dela
+
 function removeAccount(id) {
   let usuariosCadastrados = JSON.parse(localStorage.getItem("dadosCadastros"))
 
-  let index = usuariosCadastrados.find(function(usuariosCadastrados){
-      return usuariosCadastrados.id === id 
-  });
-  usuariosCadastrados.splice(index, 1)
-  localStorage.setItem("dadosCadastros", JSON.stringify(usuariosCadastrados)) 
-  listAccounts()
+  Swal.fire({
+    title: 'Você tem certeza?',
+    text: "Você não será capaz de reverter essa ação!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonText: 'Cancelar',
+    confirmButtonText: 'Sim, excluir'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        'Excluído!',
+        'A conta foi excluída.',
+        'success'
+        )
+        let index = usuariosCadastrados.find(function(usuariosCadastrados){
+          return usuariosCadastrados.id === id 
+        });
+        usuariosCadastrados.splice(index, 1)
+        localStorage.setItem("dadosCadastros", JSON.stringify(usuariosCadastrados)) 
+        listAccounts()
+    }
+  })
 }
 
-function listcollectionPoints() {
+// função que lista os pontos de coleta cadastrados no sistema. pega os dados do localStorage e mostra em forma de tabela
+
+function listCollectionPoints() {
   let collectionPoint = JSON.parse(localStorage.getItem("collectionPoint"))
   let htmlListString = `<table>
               <thead>
@@ -89,7 +112,7 @@ function listcollectionPoints() {
             <td class="cell100 column1">${i}</td>
             <td class="cell100 column2">${collectionPoint[i].cep}, ${collectionPoint[i].localidade}, ${collectionPoint[i].bairro},\n${collectionPoint[i].logradouro}, ${collectionPoint[i].numero}, ${collectionPoint[i].complemento}</td>
             <td class="cell100 column5">${collectionPoint[i].acceptedTrash}</td>
-            <td class="cell100 column1"><button class='btn btn-danger' onclick="removecollectionPoint(${i})">Excluir</button></td>
+            <td class="cell100 column1"><button class='btn btn-danger' onclick="removeCollectionPoint(${i})">Excluir</button></td>
         </tr>
     </tbody>`
 
@@ -99,17 +122,37 @@ function listcollectionPoints() {
   document.getElementById('list').innerHTML = htmlListString
 }
 
-function removecollectionPoint(id) {
+// função que exclui os pontos de coleta pelo ID
+
+function removeCollectionPoint(id) {
   let collectionPoint = JSON.parse(localStorage.getItem("collectionPoint"))
 
-  let index = collectionPoint.find(function(collectionPoint){
-      return collectionPoint.id === id 
-  });
-  collectionPoint.splice(index, 1)
-  localStorage.setItem("collectionPoint", JSON.stringify(collectionPoint)) 
-  listcollectionPoints()
+  Swal.fire({
+    title: 'Você tem certeza?',
+    text: "Você não será capaz de reverter essa ação!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonText: 'Cancelar',
+    confirmButtonText: 'Sim, excluir'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        'Excluído!',
+        'O ponto de coleta foi excluído.',
+        'success'
+        )
+        let index = collectionPoint.find(function(collectionPoint){
+          return collectionPoint.id === id 
+        });
+        collectionPoint.splice(index, 1)
+        localStorage.setItem("collectionPoint", JSON.stringify(collectionPoint)) 
+        listCollectionPoints()
+    }
+  })
 }
 
+// função que lista os estoques de lixo cadastrados no sistema. pega os dados do localStorage e mostra em forma de tabela
 
 function listStockPoints() {
   let stockPoints = JSON.parse(localStorage.getItem("stockPoints"))
@@ -145,7 +188,7 @@ function listStockPoints() {
         <tr class="row100 body">
             <td class="cell100 column1">${i}</td>
             <td class="cell100 column2">${stockPoints[i].cep}, ${stockPoints[i].localidade}, ${stockPoints[i].bairro},\n${stockPoints[i].logradouro}, ${stockPoints[i].numero}, ${stockPoints[i].complemento}</td>
-            <td class="cell100 column5">${newPointItems[i].trashType}</td>
+            <td class="cell100 column5">${stockPoints[i].stockItems}</td>
             <td class="cell100 column1"><button class='btn btn-danger' onclick="removeStockPoint(${i})">Excluir</button></td>
         </tr>
     </tbody>`
@@ -156,13 +199,32 @@ function listStockPoints() {
   document.getElementById('list').innerHTML = htmlListString
 }
 
+// função que exclui os estoques de lixo pelo ID
+
 function removeStockPoint(id) {
   let stockPoints = JSON.parse(localStorage.getItem("stockPoints"))
 
-  let index = stockPoints.find(function(stockPoints){
-      return stockPoints.id === id 
-  });
-  stockPoints.splice(index, 1)
-  localStorage.setItem("stockPoints", JSON.stringify(stockPoints)) 
-  listStockPoints()
+  Swal.fire({
+      title: 'Você tem certeza?',
+      text: "Você não será capaz de reverter essa ação!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Sim, excluir'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Excluído!',
+          'O estoque de lixo foi excluído.',
+          'success'
+          )
+          let index = stockPoints.find(function(stockPoint){
+              return stockPoint.id === id 
+          });
+          stockPoints.splice(index, 1)
+          localStorage.setItem("stockPoints", JSON.stringify(stockPoints)) 
+          listStockPoints()
+      }
+    })
 }
